+++ src/plugins/acct_gather_profile/hdf5/acct_gather_profile_hdf5.c
@@ -464,7 +464,7 @@ extern int acct_gather_profile_p_task_end(pid_t taskpid)
 	return SLURM_SUCCESS;
 }
 
-extern int acct_gather_profile_p_create_group(const char* name)
+extern int64_t acct_gather_profile_p_create_group(const char* name)
 {
 	hid_t gid_group = make_group(gid_node, name);
 	if (gid_group < 0) {
@@ -480,7 +480,8 @@ extern int acct_gather_profile_p_create_group(const char* name)
 }
 
 extern int acct_gather_profile_p_create_dataset(
-	const char* name, int parent, acct_gather_profile_dataset_t *dataset)
+	const char* name, int64_t parent,
+	acct_gather_profile_dataset_t *dataset)
 {
 	size_t type_size;
 	size_t offset, field_size;
